Apache URL Rewrite
Added .htaccess file that rewrites all URLs that are not located in the directory structure to the root directory /api.php.

Will be adding the api.php script soon. That will be the controller. So for future development this will allow us to make calls like degreepath.com/api/course/1 and that will give us the course with course id 1.