Updated Course Info API Call to REST
To get course info, just make the api call for the courses resource, with the id of the course you want information on.

ex.
/courses/1
IF you use a get and use the above resource path, you will get the info for the user.
No need to have /info if you use true RESTful design.