Changed htaccess rules and made paths easier So, now all your require_once paths start with the folder that the file is in relative to the root directory of our application. So, if you are doing something from index.php, you include api/api.php, and if you are doing something in a folder called say api/new_api.php, to include the prefix file you say require_once 'api/prefix.php' instead of just 'prefix.php'.
I also changed some of the configured directories for slim and smarty to reflect this.

.htaccess was added to root directory and changed to allow things to happn like in index.php.
Basically any page that is in the root folder that is a php file will just run as is.
You don't need to use the format for the $app->get..
All you have to do is include api/api.php for any file that you make in the root directory of our application.

So now you should be able to go to localhost and actually load the default page and it should work.
If it does not, email me. I will help you fix it.