Users API Complete.
POST /users to create a user account.
DELETE /users to remove an account.
GET /user to see the currently logged in user.
POST /user to login a user.
DELETE /user to logout a user.

The above functionality is enough to actually integrate users into the UI, but I will still be updating the API.

Changes to Come:
PUT /user to update a users's account

[#66179720]