Users Update 1
Changed Index.php page to check if the user is logged in and redirect accordingly.
Made the schedule page the default page they see when they are logged in.
Updated the ajax calls to use the current logged in user, and updated the PHP call in schedule.php to also use the current logged in user to render the schedule shown.