DON'T pull this branch or commit. UNSTABLE
Only wanted to push it so you could see where I am.
DELETE and PUT are proviing to be a PAIN!

In a bit I may abandon that whole route and just make some stupid API for it.
That will come later though, but I'll try to have it done by 7PM.