Fixes #11: Created id, and composite primary key on user_id and id.
So, this is actually what I meant to do in the first place :)
Sorry about the confusion.

So, on one side, Naomi, you were right. It seems there has to be some unique key for every table.
What I was trying to, and what I thought was possible, was just skip this part and just have a clustered index on the table, but MySQL doesn't do what I was thinking I guess, so that is why I got confused when reading instructions online.

Anyway, now what happens is each row can be identified by an id, user_id, and the term code of the semester.
However, when you actually look at the table at any given time, it will always be physically sorted by user_id.
Something like this.
user_id     id          term_code
2      	    4           ....
2      	    5           ....
2           6           ...
2           7           ..
5           1           .
5           2           .
5           3	        .
7           8	        .
7           9	        .
7           10	        .
7           11	        .
7           12	        .
7           13	        .

So that way when we make queries, they are acutally sorted pysically in space by user_id.
Now whether this will make a HUGE difference in queries is debable, but if we want this website to scale to a lot of users, I think this is something we should do in our tables. I'm going to go through and do this to some other tables that use user_id.